Added pom.xml and moved functionality to Proquint class
