+++ pom.xml
@@ -1,0 +1,63 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.sonatype.oss</groupId>
+    <artifactId>oss-parent</artifactId>
+    <version>7</version>
+  </parent>
+  <groupId>com.github.dsw.proquint</groupId>
+  <artifactId>proquint</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
+  <name>Generate identifiers that are Readable, Spellable, and Pronounceable</name>
+  <developers>
+    <developer>
+      <id>dsw</id>
+      <name>Daniel S. Wilkerson</name>
+      <email>daniel.wilkerson@gmail.com</email>
+    </developer>
+    <developer>
+      <id>mwanji</id>
+      <name>Moandji Ezana</name>
+      <email>mwanji@gmail.com</email>
+    </developer>
+  </developers>
+  <licenses>
+    <license>
+      <name>
+        BSD 2-clause &quot;Simplified&quot; or &quot;FreeBSD&quot; License
+      </name>
+      <url>http://opensource.org/licenses/BSD-2-Clause</url>
+    </license>
+  </licenses>
+  <url>http://github.com/dsw/proquint</url>
+  <issueManagement>
+    <system>Github Issues</system>
+    <url>http://github.com/dsw/proquint/issues</url>
+  </issueManagement>
+  <scm>
+    <connection>scm:git:git@github.com:dsw/proquint</connection>
+    <developerConnection>scm:git:git@github.com:dsw/proquint</developerConnection>
+    <url>https://github.com/dsw/proquint</url>
+  </scm>
+  <dependencies>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.11</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.1</version>
+        <configuration>
+          <source>1.5</source>
+          <target>1.5</target>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+</project>
